feat: connect NewProfileImage and NewNickname
